Corrected a display bug on recommending a script to call. Added code in the undercloud VM creation to support existing network bridges or even create the network bridges for the option to use physical NIC's as opposed to an all virtual network config that is used in a lab/demo environment.

+++ undercloud_configure.sh
@@ -12,10 +12,37 @@ stdout "This script will gather the data for the undercloud deployment and insta
 stdout "necessary packages to do the undercloud install."
 stdout ""
 
+stdout ""
+stdout "Please confirm the following looks correct:"
+stdout ""
+stdout "Undercloud VM IP: ${UNDERCLOUD_IP}"
+stdout "Undercloud user: ${UNDERCLOUD_USER}"
+stdout ""
+
+USER_INPUT=''
+
+while [[ -z "${USER_INPUT}" ]]
+do
+  read -p "Are these values correct [Y/n]? " USER_INPUT
+  if [[ -z "${USER_INPUT}" ]]
+  then
+    USER_INPUT='Y'
+  fi
+  USER_INPUT=$(echo ${USER_INPUT} | tr '[:lower:]' '[:upper:]' | cut -c1 | egrep '(Y|N)')
+done
+
+if [[ "${USER_INPUT}" == 'N' ]]
+then
+  stdout ""
+  stdout "Update the settings in ${DIRECTOR_TOOLS}/environment/undercloud.env and then re-run this script."
+  stdout ""
+  exit 2
+fi
+
 rm -f ${DIRECTOR_TOOLS}/logs/*undercloud_configure*log*
 ${DIRECTOR_TOOLS}/functions/undercloud/configure.sh
 
-if [[ ! -z "$(ls ${DIRECTOR_TOOLS}/logs/*undercloud_configure*err*)" ]]
+if [[ ! -z "$(ls ${DIRECTOR_TOOLS}/logs/*undercloud_configure*err* 2>/dev/null)" ]]
 then
   stderr "There was an error configuring the undercloud."
   exit 300
